Reorganize client code again
