+++ program/src/utils.rs
@@ -8,36 +8,6 @@ use spl_token::{error::TokenError, state::Account as TokenAccount};
 use thiserror::Error;
 use std::str::FromStr;
 
-pub fn program_id() -> Pubkey {
-    Pubkey::from_str("Fpwgc9Tq7k2nMzVxYqPWwKGA7FbCQwo2BgekpT69Cgbf").unwrap()
-}
-
-pub fn get_counter_address(payee: &Pubkey, amount: u64, duration: i64) -> (Pubkey, u8) {
-    let counter_seeds = &[
-        b"subscription_counter",
-        payee.as_ref(),
-        &amount.to_le_bytes(),
-        &duration.to_le_bytes(),
-    ];
-    Pubkey::find_program_address(counter_seeds, &program_id())
-}
-
-pub fn get_subscription_address(
-    payee: &Pubkey,
-    amount: u64,
-    duration: i64,
-    count: u64,
-) -> (Pubkey, u8) {
-    let subscription_seeds = &[
-        b"subscription_metadata",
-        payee.as_ref(),
-        &amount.to_le_bytes(),
-        &duration.to_le_bytes(),
-        &count.to_le_bytes(),
-    ];
-    Pubkey::find_program_address(subscription_seeds, &program_id())
-}
-
 // ACCOUNT VALIDATION
 
 pub fn assert_msg(statement: bool, err: ProgramError, msg: &str) -> ProgramResult {
